[fix]components以下のディレクトリ名とpages以下のディレクトリ名を同一に変更(#280)
