+++ view/next-project/src/components/budgets/OpenDeleteModalButton.tsx
@@ -1,24 +1,18 @@
 import { ChakraProvider, Button } from '@chakra-ui/react';
 import theme from '@assets/theme';
 import * as React from 'react';
-import EditModal from '@components/user/EditModal';
 import { useState } from 'react';
-import { RiPencilFill } from 'react-icons/ri';
-
-interface Bureau {
-  id: number;
-  name: string;
-}
+import { RiDeleteBinLine } from 'react-icons/ri';
+import DeleteModal from '@components/budgets/DeleteModal';
 
 interface Props {
   width?: string;
   height?: string;
   children?: React.ReactNode;
   id: number;
-  bureaus: Bureau[];
 }
 
-const OpenEditModalButton: React.FC<Props> = (props) => {
+const OpenDeleteModalButton: React.FC<Props> = (props) => {
   const [showModal, setShowModal] = useState(false);
   const ShowModal = () => {
     setShowModal(true);
@@ -31,20 +25,15 @@ const OpenEditModalButton: React.FC<Props> = (props) => {
         p='0'
         minWidth='0'
         borderRadius='full'
-        bgGradient='linear(to-br, primary.1 ,primary.2)'
+        bgGradient='linear(to-br, red.500 ,red.600)'
         onClick={ShowModal}
       >
-        <RiPencilFill size={'15px'} color={'white'} />
+        <RiDeleteBinLine size={'15px'} color={'white'} />
         {props.children}
       </Button>
-      <EditModal
-        id={props.id}
-        bureaus={props.bureaus}
-        openModal={showModal}
-        setShowModal={setShowModal}
-      />
+      <DeleteModal id={props.id} openModal={showModal} setShowModal={setShowModal} />
     </ChakraProvider>
   );
 };
 
-export default OpenEditModalButton;
+export default OpenDeleteModalButton;
