+++ view/next-project/src/components/fund_information/DetailModal.tsx
