+++ view/next-project/src/components/purchaseorders/OpenEditModalButton.tsx
@@ -1,9 +1,9 @@
 import { ChakraProvider, Button } from '@chakra-ui/react';
 import theme from '@assets/theme';
 import * as React from 'react';
-import PurchaseOrderDeleteModal from '@components/purchaseorder/PurchaseOrderDeleteModal';
+import EditModal from '@components/purchaseorders/EditModal';
 import { useState } from 'react';
-import { RiDeleteBinLine } from 'react-icons/ri';
+import { RiPencilFill } from 'react-icons/ri';
 
 interface Props {
   width?: string;
@@ -12,7 +12,7 @@ interface Props {
   id: number;
 }
 
-const OpenDeleteModalButton: React.FC<Props> = (props) => {
+const OpenEditModalButton: React.FC<Props> = (props) => {
   const [showModal, setShowModal] = useState(false);
   const ShowModal = () => {
     setShowModal(true);
@@ -28,12 +28,12 @@ const OpenDeleteModalButton: React.FC<Props> = (props) => {
         bgGradient='linear(to-br, primary.1 ,primary.2)'
         onClick={ShowModal}
       >
-        <RiDeleteBinLine size={'15px'} color={'white'} />
+        <RiPencilFill size={'15px'} color={'white'} />
         {props.children}
       </Button>
-      <PurchaseOrderDeleteModal id={props.id} openModal={showModal} setShowModal={setShowModal} />
+      <EditModal id={props.id} openModal={showModal} setShowModal={setShowModal} />
     </ChakraProvider>
   );
 };
 
-export default OpenDeleteModalButton;
+export default OpenEditModalButton;
