+++ view/next-project/src/components/purchasereports/OpenEditModalButton.tsx
@@ -1,9 +1,9 @@
 import { ChakraProvider, Button } from '@chakra-ui/react';
 import theme from '@assets/theme';
 import * as React from 'react';
+import EditModal from '@components/purchasereports/EditModal';
 import { useState } from 'react';
-import { RiDeleteBinLine } from 'react-icons/ri';
-import BudgetDeleteModal from '@components/budget/BudgetDeleteModal';
+import { RiPencilFill } from 'react-icons/ri';
 
 interface Props {
   width?: string;
@@ -12,7 +12,7 @@ interface Props {
   id: number;
 }
 
-const OpenDeleteModalButton: React.FC<Props> = (props) => {
+const OpenEditModalButton: React.FC<Props> = (props) => {
   const [showModal, setShowModal] = useState(false);
   const ShowModal = () => {
     setShowModal(true);
@@ -25,15 +25,15 @@ const OpenDeleteModalButton: React.FC<Props> = (props) => {
         p='0'
         minWidth='0'
         borderRadius='full'
-        bgGradient='linear(to-br, red.500 ,red.600)'
+        bgGradient='linear(to-br, primary.1 ,primary.2)'
         onClick={ShowModal}
       >
-        <RiDeleteBinLine size={'15px'} color={'white'} />
+        <RiPencilFill size={'15px'} color={'white'} />
         {props.children}
       </Button>
-      <BudgetDeleteModal id={props.id} openModal={showModal} setShowModal={setShowModal} />
+      <EditModal id={props.id} openModal={showModal} setShowModal={setShowModal} />
     </ChakraProvider>
   );
 };
 
-export default OpenDeleteModalButton;
+export default OpenEditModalButton;
