+++ view/next-project/src/components/teachers/DeleteModal.tsx
