+++ view/next-project/src/components/teachers/DetailModal.tsx
